+++ src/App/Header/index.tsx
@@ -1,13 +1,11 @@
-import Image, { DappLogo } from "components/Image";
-import QueryLoader from "components/QueryLoader";
+import { DappLogo } from "components/Image";
 import { appRoutes } from "constants/routes";
-import { categories } from "constants/unsdgs";
-import useDebouncer from "hooks/useDebouncer";
 import { useState } from "react";
-import { Link, useLocation } from "react-router-dom";
-import { useEndowmentCardsQuery } from "services/aws/aws";
+import { useLocation } from "react-router-dom";
+
 import type { Link as TLink } from "../types";
 import NavDropdown from "./NavDropdown";
+import SearchDropdown from "./SearchDropdown";
 import SearchField from "./SearchField";
 import UserMenu from "./UserMenu";
 
@@ -16,21 +14,6 @@ type Props = { links: TLink[]; classes?: string };
 export default function Header({ links, classes }: Props) {
   const location = useLocation();
   const [query, setQuery] = useState("");
-  const [debouncedQuery, isDebouncing] = useDebouncer(query, 500);
-
-  const { currentData, isLoading, isFetching, isError, isUninitialized } =
-    useEndowmentCardsQuery(
-      {
-        query: debouncedQuery,
-        page: 1,
-      },
-      { skip: isDebouncing }
-    );
-
-  const isCategoriesShown =
-    !query && !isDebouncing && !isLoading && !isFetching;
-
-  const hasMoreItems = (currentData?.NumOfPages ?? 1) > 1;
 
   return (
     <header
@@ -69,68 +52,10 @@ export default function Header({ links, classes }: Props) {
         </div>
       </div>
       {location.pathname === appRoutes.home && (
-        <div className="absolute hidden group-has-[input:focus]:block hover:block bg-white mt-4 container left-1/2 -translate-x-1/2 rounded-lg p-6 shadow-2xl shadow-black/20">
-          {isCategoriesShown ? (
-            <>
-              <h4 className="mb-4">Top categories</h4>
-              <div className="flex flex-wrap gap-2">
-                {Object.values(categories).map((v) => (
-                  <Link
-                    key={v.name}
-                    className="border border-blue-d1 px-6 py-2 rounded-full text-sm"
-                    to={appRoutes.marketplace}
-                  >
-                    {v.name}
-                  </Link>
-                ))}
-              </div>
-              <Link
-                className="text-blue-d1 font-medium text-lg text-center mt-8 block"
-                to={appRoutes.marketplace}
-              >
-                Explore all Categories
-              </Link>
-            </>
-          ) : (
-            <QueryLoader
-              queryState={{
-                data: currentData?.Items || [],
-                isLoading: isLoading || isUninitialized,
-                isFetching: isFetching || isUninitialized,
-                isError,
-              }}
-              messages={{
-                loading: "Searching...",
-                fetching: "Searching...",
-                empty: "No endowments found",
-              }}
-            >
-              {(endows) => (
-                <div className="flex flex-wrap items-center gap-4">
-                  {endows.map((endow) => (
-                    <Link
-                      to={`${appRoutes.marketplace}/${endow.id}`}
-                      key={endow.id}
-                      className="flex items-center gap-4 border border-gray-l4 pr-6 rounded-full overflow-clip"
-                    >
-                      <Image
-                        src={endow.card_img}
-                        className="h-10 w-14 object-contain"
-                      />
-                      <span>{endow.name}</span>
-                    </Link>
-                  ))}
-                  <Link
-                    className="w-full text-blue-d1 font-medium text-lg text-center mt-8 block"
-                    to={appRoutes.marketplace}
-                  >
-                    View all results
-                  </Link>
-                </div>
-              )}
-            </QueryLoader>
-          )}
-        </div>
+        <SearchDropdown
+          query={query}
+          classes="mt-4 hidden group-has-[input:focus]:block hover:block absolute left-1/2 -translate-x-1/2"
+        />
       )}
     </header>
   );
